+++ src/datalevin/core.clj
@@ -2219,9 +2219,11 @@ formally with the following grammar:
     <expression> ::= <term> | [ <operator> <operands> ]
     <operator>   ::= :or | :and | :not
     <operands>   ::= <expression>+
-    <term>       ::= string
+    <term>       ::= string | { <pair>+ }
+    <pair>       ::= <key> string
+    <key>        ::= :phrase | :term
 
-For example, [:or [:and \"red\" \"fox\" [:not \"lazy\"]] \"jump\"]
+For example, [:or [:and \"red\" \"fox\" [:not \"lazy\"]] {:phrase \"jump over\"}]
 
 If the query is a string of words, e.g. \"word1 word2 word3\", it is equivalent to
 [:or \"word1\" \"word2\" \"word3\"] when using the default analyzer.