+++ test/datalevin/search_test.clj
@@ -50,6 +50,7 @@
                         (sut/parse-query* a/en-analyzer query))
     []
     [:or]
+    [:not]
     ["book"]
     ["book" "red"]
     [:none "book"]))
@@ -118,6 +119,8 @@
       [:or "nonexistentterm" "whale"]       #{:doc3}
       [:not "red"]                          #{:doc3}
       [:not [:and "red" "fox"]]             #{:doc2 :doc3 :doc4 :doc5}
+      [:not "red fox"]                      #{:doc3}
+      "red fox"                             #{:doc1 :doc2 :doc4 :doc5}
 
       [:or "fox" "red" [:and "black" "sheep"] [:not "yellow"]]
       #{:doc1 :doc2 :doc4 :doc5}
@@ -132,10 +135,48 @@
       "nonexistentterm"
       [:and "nonexistentterm" "red"]
       [:and "red" [:not "red"]] ; Conflicting condition
-      [:or "" ""])
+      [:or "" ""]
+      )
 
     (l/close-kv lmdb)
     (u/delete-files dir)))
+
+#_(deftest phrase-test
+    (let [dir    (u/tmp-dir (str "test-" (UUID/randomUUID)))
+          lmdb   (l/open-kv dir)
+          engine (sut/new-search-engine lmdb {:index-position? true})]
+      (add-docs sut/add-doc engine)
+
+      (are [query result] (= result (set (sut/search engine query)))
+        [:and "red" [:not {:phrase "red fox"}]] #{:doc1 :doc5}
+        ;; [:and "red" [:not "nonexistentterm"]] #{:doc1 :doc2 :doc4 :doc5}
+        ;; [:and "man" "whale"]                  #{:doc3}
+        ;; [:and "red" [:not "dogs"]]            #{:doc2 :doc4}
+        ;; [:or "" "whale"]                      #{:doc3}
+        ;; [:or "nonexistentterm" "whale"]       #{:doc3}
+        ;; [:not "red"]                          #{:doc3}
+        ;; [:not [:and "red" "fox"]]             #{:doc2 :doc3 :doc4 :doc5}
+        ;; [:not "red fox"]                      #{:doc3}
+        ;; "red fox"                             #{:doc1 :doc2 :doc4 :doc5}
+
+        ;; [:or "fox" "red" [:and "black" "sheep"] [:not "yellow"]]
+        ;; #{:doc1 :doc2 :doc4 :doc5}
+
+        ;; [:or "fox" "red" [:and "black" "sheep"] [:not "fleece"]]
+        ;; #{:doc1 :doc2 :doc4 :doc5}
+        )
+
+      #_(are [query] (empty? (sut/search engine query))
+          ;; ""
+          ;; "  "
+          ;; "nonexistentterm"
+          ;; [:and "nonexistentterm" "red"]
+          ;; [:and "red" [:not "red"]] ; Conflicting condition
+          ;; [:or "" ""]
+          )
+
+      (l/close-kv lmdb)
+      (u/delete-files dir)))
 
 (deftest index-test
   (let [dir           (u/tmp-dir (str "index-" (UUID/randomUUID)))